Improves application setup and adds UI framework
This commit enhances the initial application setup for a better developer experience.

It configures the application for Indonesian localization, switches to cookie-based sessions, and sets the timezone to Asia/Jakarta.

Additionally, this commit integrates Livewire, Livewire-Flux and spatie/laravel-permission for UI components and role-based access control.
A debugbar is also added for development purposes.